before first merge
