+++ src/main/java/com/example/demo/controller/HomeController.java
@@ -1,0 +1,15 @@
+package com.example.demo.controller;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+@AllArgsConstructor
+public class HomeController {
+
+    @GetMapping("/home")
+    public String dispHome(){
+        return "index";
+    }
+}
