+++ src/main/java/com/example/demo/repository/PaymentRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.domain.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface PaymentRepository extends JpaRepository<Account, Long> {
 }
