+++ src/main/java/com/example/demo/service/MemberService.java
@@ -6,6 +6,7 @@ import com.example.demo.dto.MemberDto;
 import com.example.demo.repository.MemberRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -16,6 +17,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
@@ -47,4 +49,6 @@ public class MemberService implements UserDetailsService {
         }
         return new User(userEntity.getEmail(),userEntity.getPassword(),authorities);
     }
+
+
 }
